+++ server/src/api/auth/index.ts
