+++ server/src/app.ts
@@ -5,9 +5,9 @@ import * as session from "express-session";
 import * as cookieParser from "cookie-parser";
 import * as passport from "passport";
 import * as cors from "cors";
+import apiRouter from "./api/auth";
+import passportConfig from "./passport";
 
-import apiRouter from "./api/routes/auth";
-import passportConfig from "./api/passport";
 dotenv.config();
 
 const app: express.Application = express();
